Final changes made for background color and the folders organization, carrusel was moved onto components instead of being in screens
