+++ Desarrollo2/src/components/Carrusel/carruselInfinito.css
