+++ Desarrollo2/src/screens/mainScreen/principal.css
@@ -2,7 +2,7 @@ body {
     margin: 0;
     padding: 0;
     font-family: "Arial", sans-serif;
-    background: linear-gradient(135deg, #667eea, #764ba2);
+    background: linear-gradient(135deg, #ee0e0e, #b659cd);
     min-height: 100vh;
 }
